MDL-73941 reportbuilder: return localised floats when aggregating.
As the 'average' and 'percentage' aggregation types both return
float values, we should ensure they are output with localised
formatting (e.g. decimal separator).

+++ reportbuilder/classes/local/aggregation/percent.php
@@ -72,6 +72,6 @@ class percent extends base {
      * @return mixed
      */
     public static function format_value($value, array $values, array $callbacks) {
-        return format::percent((float) reset($values));
+        return format::percent(reset($values));
     }
 }
